requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.2
requests==2.33.1
telebot==0.0.5
urllib3==2.6.3
